refactor: productInfo 데이터 클래스 위치 dto로 이동
