+++ android/app/src/main/java/app/priceguard/data/dto/ProductInfo.kt
@@ -1,0 +1,8 @@
+package app.priceguard.data.dto
+
+data class ProductInfo(
+    val title: String = "",
+    val brand: String = "",
+    val logo: Int = 0,
+    val price: String = ""
+)
